Fixed types

+++ inventory/addItemBulk.go
@@ -6,10 +6,10 @@ type AddItemBulk struct {
 		Code         string `json:"Code"`
 		LocationCode string `json:"LocationCode"`
 		Note         string `json:"Note"`
-		Quantity     int64  `json:"Quantity"`
+		Quantity     int  `json:"Quantity"`
 		Reason       string `json:"Reason"`
 		Sku          string `json:"Sku"`
-		WarehouseID  int64  `json:"WarehouseId"`
+		WarehouseID  int  `json:"WarehouseId"`
 	} `json:"Items"`
 }
