+++ inventory/getWarehouses.go
@@ -2,7 +2,7 @@ package inventory
 
 // GetWarehouses is a automatically generated struct from json provided by sku vault's api docs.
 type GetWarehouses struct {
-	PageNumber int64 `json:"PageNumber"`
+	PageNumber                int `json:"PageNumber"`
 }
 
 // GetWarehousesResponse is a automatically generated struct from json provided by sku vault's api docs.
