+++ inventory/pickItem.go
@@ -6,9 +6,9 @@ type PickItem struct {
 	IsExpressPick bool   `json:"IsExpressPick"`
 	LocationCode  string `json:"LocationCode"`
 	Note          string `json:"Note"`
-	Quantity      int64  `json:"Quantity"`
+	Quantity     int  `json:"Quantity"`
 	Sku           string `json:"Sku"`
-	WarehouseID   int64  `json:"WarehouseId"`
+	WarehouseID  int  `json:"WarehouseId"`
 }
 
 // PickItemResponse is a automatically generated struct from json provided by sku vault's api docs.
