+++ inventory/pickItemBulk.go
@@ -7,11 +7,11 @@ type PickItemBulk struct {
 		IsExpressPick bool   `json:"IsExpressPick"`
 		LocationCode  string `json:"LocationCode"`
 		Note          string `json:"Note"`
-		Quantity      int64  `json:"Quantity"`
+		Quantity     int  `json:"Quantity"`
 		SaleID        string `json:"SaleId"`
 		ScannedCode   string `json:"ScannedCode"`
 		Sku           string `json:"Sku"`
-		WarehouseID   int64  `json:"WarehouseId"`
+		WarehouseID  int  `json:"WarehouseId"`
 	} `json:"Items"`
 }
