+++ inventory/updateExternalWarehouseQuantities.go
@@ -3,12 +3,12 @@ package inventory
 // UpdateExternalWarehouseQuantities is a automatically generated struct from json provided by sku vault's api docs.
 type UpdateExternalWarehouseQuantities struct {
 	Quantities []struct {
-		InStockQuantity  int64  `json:"InStockQuantity"`
-		InboundQuantity  int64  `json:"InboundQuantity"`
-		ReserveQuantity  int64  `json:"ReserveQuantity"`
+		InStockQuantity     int  `json:"InStockQuantity"`
+		InboundQuantity     int  `json:"InboundQuantity"`
+		ReserveQuantity     int  `json:"ReserveQuantity"`
 		Sku              string `json:"Sku"`
-		TotalQuantity    int64  `json:"TotalQuantity"`
-		TransferQuantity int64  `json:"TransferQuantity"`
+		TotalQuantity     int  `json:"TotalQuantity"`
+		TransferQuantity     int  `json:"TransferQuantity"`
 	} `json:"Quantities"`
 	WarehouseID string `json:"WarehouseId"`
 }
