+++ products/getSuppliers.go
@@ -2,7 +2,7 @@ package products
 
 // GetSuppliers is a automatically generated struct from json provided by sku vault's api docs.
 type GetSuppliers struct {
-	PageNumber int64 `json:"PageNumber"`
+	PageNumber                int `json:"PageNumber"`
 }
 
 // GetSuppliersResponse is a automatically generated struct from json provided by sku vault's api docs.
